+++ main.ts
@@ -1,13 +1,13 @@
-import AutoLaunch from 'auto-launch';
-import { app, BrowserWindow, dialog, Display, ipcMain, Menu, MenuItem, MenuItemConstructorOptions, screen, systemPreferences, Tray } from 'electron';
-import * as log from 'electron-log';
-import { autoUpdater } from 'electron-updater';
-import * as fs from 'fs';
-import { Window } from 'node-window-manager';
+import { app, BrowserWindow, dialog, Display, ipcMain, Menu, MenuItemConstructorOptions, screen, systemPreferences, Tray } from 'electron';
 import * as path from 'path';
-import * as robot from 'robotjs';
 import * as url from 'url';
-import * as hook from './hook';
+import * as launch from './electron/auto-launch';
+import * as update from './electron/auto-updater';
+import * as game from './electron/game';
+import * as hook from './electron/hook';
+import * as log from './electron/log';
+import * as robot from './electron/robot';
+import { State } from './electron/state';
 
 if (!app.requestSingleInstanceLock()) {
     app.exit();
@@ -25,63 +25,28 @@ app.allowRendererProcessReuse = false;
 app.commandLine.appendSwitch('high-dpi-support', 'true');
 app.commandLine.appendSwitch('force-device-scale-factor', '1');
 
-log.transports.file.level = 'info';
-Object.assign(console, log.functions);
+log.register(ipcMain);
 
-log.info('App starting...');
+console.info('App starting...');
 
-let animationPath = path.join(app.getPath('userData'), 'animation.flag');
-let animationExists = fs.existsSync(animationPath);
-
-log.info(`App checking animation flag: ${animationExists}.`);
-
-if (animationExists) {
+const state = new State(app.getPath('userData'));
+if (!state.hardwareAcceleration) {
     app.disableHardwareAcceleration();
-    log.info('App started with disabled hardware acceleration.');
+    console.info('App started with disabled hardware acceleration.');
 }
-
-let keyboardPath = path.join(app.getPath('userData'), 'keyboard.flag');
-let keyboardExists = fs.existsSync(keyboardPath);
-
-log.info(`App checking keyboard flag: ${keyboardExists}.`);
-
-let versionPath = path.join(app.getPath('userData'), 'version.txt');
-let versionExists = fs.existsSync(versionPath);
-
-let versionUpdated = true;
-if (versionExists) {
-    const version = fs.readFileSync(versionPath, 'utf-8').trim();
-    versionUpdated = version !== app.getVersion();
-    log.info(`App checking version: ${version} -> ${app.getVersion()}, ${versionUpdated}`);
-}
-if (versionUpdated) {
-    fs.writeFileSync(versionPath, app.getVersion())
-}
-
-autoUpdater.logger = log;
-autoUpdater.autoInstallOnAppQuit = true;
 
 const args = process.argv.slice(1);
+console.info('App args', args);
+
 const serve = args.some(val => val === '--serve');
-
-log.info('App args', args);
-log.info('App served', serve);
-
-const launch = new AutoLaunch({
-    name: 'PoE Overlay'
-});
 
 let win: BrowserWindow = null;
 let tray: Tray = null;
 let menu: Menu = null;
-let downloadItem: MenuItem = null;
-let checkForUpdatesHandle = null;
-let gameWindow: Window = null;
 
 const childs: {
     [key: string]: BrowserWindow
 } = {};
-
 
 /* helper */
 
@@ -94,182 +59,47 @@ function send(channel: string, ...args: any[]) {
         win.webContents.send(channel, ...args);
     }
     catch (error) {
-        log.error(`could not send to '${channel}' with args '${JSON.stringify(args)}`);
+        console.error(`could not send to '${channel}' with args '${JSON.stringify(args)}`);
     }
 }
 
-/* log */
+launch.register(ipcMain);
 
-ipcMain.on('log', (event, level, message, ...args) => {
-    log[level](message, ...args);
-    event.returnValue = true;
+update.register(ipcMain, menu, tray, event => {
+    send(event);
 });
 
-/* robot js */
+robot.register(ipcMain);
 
-ipcMain.on('click-at', (event, button, position) => {
-    if (position) {
-        robot.moveMouse(position.x, position.y);
-    }
-    robot.mouseClick(button, false);
-    event.returnValue = true;
-});
+game.register(ipcMain, poe => {
+    send('game-active-change', serve ? true : poe.active);
 
-ipcMain.on('mouse-pos', (event) => {
-    event.returnValue = robot.getMousePos()
-});
+    if (win && poe.active) {
+        win.setAlwaysOnTop(false);
+        win.setVisibleOnAllWorkspaces(false);
 
-ipcMain.on('key-tap', (event, key, modifier) => {
-    robot.keyTap(key, modifier);
-    event.returnValue = true;
-});
+        win.setAlwaysOnTop(true, 'pop-up-menu', 1);
+        win.setVisibleOnAllWorkspaces(true);
 
-ipcMain.on('key-toggle', (event, key, down, modifier) => {
-    robot.keyToggle(key, down, modifier);
-    event.returnValue = true;
-});
-
-ipcMain.on('set-keyboard-delay', (event, delay) => {
-    robot.setKeyboardDelay(delay);
-    event.returnValue = true;
-});
-
-/* hook */
-
-ipcMain.on('force-active', event => {
-    if (keyboardExists) {
-        gameWindow?.bringToTop();
-    }
-    event.returnValue = true;
-})
-
-ipcMain.on('register-active-change', event => {
-    hook.on('change', (active, activeWindow, bounds) => {
-        gameWindow = activeWindow;
-
-        send('active-change', serve ? true : active);
-
-        if (win && active) {
-            win.setAlwaysOnTop(false);
-            win.setVisibleOnAllWorkspaces(false);
-
-            win.setAlwaysOnTop(true, 'pop-up-menu', 1);
-            win.setVisibleOnAllWorkspaces(true);
-
-            if (JSON.stringify(bounds) !== JSON.stringify(win.getBounds())) {
-                win.setBounds(bounds);
-                log.info('set bounds to: ', win.getBounds());
-            }
+        if (poe.bounds) {
+            win.setBounds(poe.bounds);
         }
-
-    });
-    event.returnValue = true;
+    }
 });
 
-ipcMain.on('register-shortcut', (event, accelerator) => {
-    switch (accelerator) {
-        case 'CmdOrCtrl + MouseWheelUp':
-        case 'CmdOrCtrl + MouseWheelDown':
-            hook.on('wheel', e => {
-                if (e.ctrlKey) {
-                    const channel = `shortcut-CmdOrCtrl + ${e.rotation === -1 ? 'MouseWheelUp' : 'MouseWheelDown'}`;
-                    send(channel);
-                }
-            });
-            break;
-    }
-    event.returnValue = true;
-});
-
-ipcMain.on('unregister-shortcut', (event, accelerator) => {
-    switch (accelerator) {
-        case 'CmdOrCtrl + MouseWheelUp':
-        case 'CmdOrCtrl + MouseWheelDown':
-            hook.off('wheel');
-            break;
-    }
-    event.returnValue = true;
+hook.register(ipcMain, event => send(event), () => {
+    dialog.showErrorBox(
+        'Iohook is required to run PoE Overlay',
+        'Iohook could not be loaded. Please make sure you have vc_redist installed and try again.');
+    app.quit();
 });
 
 /* auto-updater */
-
-autoUpdater.on('update-available', () => {
-    send('app-update-available');
-    tray?.displayBalloon({
-        iconType: 'info',
-        title: 'New update available',
-        content: 'A new update is available. Will be automatically downloaded unless otherwise specified.',
-    });
-    if (!autoUpdater.autoDownload && !downloadItem) {
-        downloadItem = new MenuItem({
-            label: 'Download Update',
-            type: 'normal',
-            click: () => {
-                autoUpdater.downloadUpdate();
-                downloadItem.enabled = false;
-            }
-        });
-        menu?.insert(2, downloadItem);
-    }
-    if (checkForUpdatesHandle) {
-        clearInterval(checkForUpdatesHandle);
-        checkForUpdatesHandle = null;
-    }
-});
-
-autoUpdater.on('update-downloaded', () => {
-    send('app-update-downloaded');
-    tray?.displayBalloon({
-        iconType: 'info',
-        title: 'Update ready to install',
-        content: 'The new update is now ready to install. Please relaunch your application.',
-    });
-});
-
-ipcMain.on('app-download-init', (event, autoDownload) => {
-    autoUpdater.autoDownload = autoDownload;
-    if (!serve) {
-        autoUpdater.checkForUpdates();
-        checkForUpdatesHandle = setInterval(() => {
-            autoUpdater.checkForUpdates();
-        }, 1000 * 60 * 5);
-    }
-    event.returnValue = true;
-});
-
-ipcMain.on('app-download-auto', (event, autoDownload) => {
-    autoUpdater.autoDownload = autoDownload;
-    event.returnValue = true;
-});
-
-ipcMain.on('app-download-update', event => {
-    autoUpdater.downloadUpdate();
-    event.returnValue = true;
-});
-
-ipcMain.on('app-quit-and-install', (event, restart) => {
-    autoUpdater.quitAndInstall(false, restart);
-    event.returnValue = true;
-});
 
 ipcMain.on('app-version', event => {
     const version = app.getVersion();
     event.returnValue = version;
 });
-
-/* auto-launch */
-
-ipcMain.on('app-auto-launch-enabled', event => {
-    launch.isEnabled()
-        .then((enabled: boolean) => event.sender.send('app-auto-launch-enabled-result', enabled))
-        .catch(() => event.sender.send('app-auto-launch-enabled-result', false));
-});
-
-ipcMain.on('app-auto-launch-change', (event, enabled) => {
-    (enabled ? launch.enable() : launch.disable())
-        .then(() => event.sender.send('app-auto-launch-change-result', true))
-        .catch(() => event.sender.send('app-auto-launch-change-result', false));
-})
 
 /* change log */
 function showChangelog() {
@@ -301,7 +131,7 @@ function createWindow(): BrowserWindow {
             allowRunningInsecureContent: serve,
             webSecurity: false
         },
-        focusable: keyboardExists,
+        focusable: state.keyboardSupport,
         skipTaskbar: true,
         show: false
     });
@@ -312,7 +142,7 @@ function createWindow(): BrowserWindow {
     win.setVisibleOnAllWorkspaces(true);
 
     win.once('show', () => {
-        if (versionUpdated) {
+        if (state.isVersionUpdated(app.getVersion())) {
             showChangelog();
         }
     })
@@ -417,23 +247,15 @@ function createTray(): Tray {
         },
         {
             label: 'Keyboard Support (experimental)', type: 'checkbox',
-            checked: keyboardExists, click: () => {
-                if (keyboardExists) {
-                    fs.unlinkSync(keyboardPath);
-                } else {
-                    fs.writeFileSync(keyboardPath, 'true');
-                }
+            checked: state.keyboardSupport, click: () => {
+                state.keyboardSupport = !state.keyboardSupport;
                 send('app-relaunch');
             }
         },
         {
             label: 'Hardware Acceleration', type: 'checkbox',
-            checked: !animationExists, click: () => {
-                if (animationExists) {
-                    fs.unlinkSync(animationPath);
-                } else {
-                    fs.writeFileSync(animationPath, 'true');
-                }
+            checked: state.hardwareAcceleration, click: () => {
+                state.hardwareAcceleration = !state.hardwareAcceleration;
                 send('app-relaunch');
             }
         },
@@ -465,17 +287,8 @@ try {
     app.on('ready', () => {
         /* delay create window in order to support transparent windows at linux. */
         setTimeout(() => {
-            hook.register().then(success => {
-                if (!success) {
-                    dialog.showErrorBox(
-                        'Iohook is required to run PoE Overlay',
-                        'Iohook could not be loaded. Please make sure you have vc_redist installed and try again.');
-                    app.quit();
-                } else {
-                    createWindow();
-                    createTray();
-                }
-            });
+            createWindow();
+            createTray();
         }, 300);
     });