+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.6.17 (2020-04-03)
 
+- update run iohook only if required (stash navigation)
 - fix quality higher than 20% showing a lower value as max
 
 ## 0.6.17 (2020-04-02)